+++ src/test/java/org/locationtech/proj4j/ExampleTest.java
@@ -15,12 +15,10 @@
  *******************************************************************************/
 package org.locationtech.proj4j;
 
+import org.junit.Assert;
 import org.junit.Test;
-import org.locationtech.proj4j.CRSFactory;
-import org.locationtech.proj4j.CoordinateReferenceSystem;
-import org.locationtech.proj4j.CoordinateTransform;
-import org.locationtech.proj4j.CoordinateTransformFactory;
-import org.locationtech.proj4j.ProjCoordinate;
+
+import java.io.IOException;
 
 import static org.junit.Assert.assertTrue;
 
@@ -41,16 +39,16 @@ public class ExampleTest {
         CoordinateTransformFactory ctFactory = new CoordinateTransformFactory();
         CRSFactory csFactory = new CRSFactory();
         /*
-        * Create {@link CoordinateReferenceSystem} & CoordinateTransformation.
-        * Normally this would be carried out once and reused for all transformations
-        */
+         * Create {@link CoordinateReferenceSystem} & CoordinateTransformation.
+         * Normally this would be carried out once and reused for all transformations
+         */
         CoordinateReferenceSystem crs = csFactory.createFromName(csName);
 
         final String WGS84_PARAM = "+title=long/lat:WGS84 +proj=longlat +ellps=WGS84 +datum=WGS84 +units=degrees";
         CoordinateReferenceSystem WGS84 = csFactory.createFromParameters("WGS84", WGS84_PARAM);
 
         CoordinateTransform trans = ctFactory.createTransform(WGS84, crs);
-    
+
         /*
          * Create input and output points.
          * These can be constructed once per thread and reused.
@@ -59,7 +57,7 @@ public class ExampleTest {
         ProjCoordinate p2 = new ProjCoordinate();
         p.x = lon;
         p.y = lat;
-    
+
         /*
          * Transform point
          */
@@ -82,7 +80,7 @@ public class ExampleTest {
         CoordinateReferenceSystem crs2 = csFactory.createFromName(csName2);
 
         CoordinateTransform trans = ctFactory.createTransform(crs1, crs2);
-    
+
         /*
          * Create input and output points.
          * These can be constructed once per thread and reused.
@@ -91,7 +89,7 @@ public class ExampleTest {
         ProjCoordinate p2 = new ProjCoordinate();
         p1.x = 500000;
         p1.y = 4649776.22482;
-    
+
         /*
          * Transform point
          */
@@ -200,6 +198,36 @@ public class ExampleTest {
         assertTrue(isInTolerance(pi, p.x, p.y, 0.000001));
     }
 
+    @Test
+    public void latLonToStereBidirectionalTransform() {
+        CRSFactory csFactory = new CRSFactory();
+        CoordinateReferenceSystem STERE0 = csFactory.createFromParameters("STERE", "+proj=stere +lat_0=0.0 +lon_0=0.0 +k=1 +x_0=0 +y_0=0 +ellps=WGS84 +units=m +no_defs");
+        CoordinateReferenceSystem STERE1 = csFactory.createFromParameters("STERE", "+proj=stere +lat_0=0.000001 +lon_0=0.0 +k=1 +x_0=0 +y_0=0 +ellps=WGS84 +units=m +no_defs");
+        CoordinateReferenceSystem WGS84 = csFactory.createFromParameters("WGS84", "+proj=latlong +datum=WGS84 +ellps=WGS84 +no_defs");
+        CoordinateTransform transformer0 = new CoordinateTransformFactory().createTransform(WGS84, STERE0);
+        CoordinateTransform transformer1 = new CoordinateTransformFactory().createTransform(WGS84, STERE1);
+
+        ProjCoordinate pc = new ProjCoordinate(1, 1);
+        ProjCoordinate result0 = new ProjCoordinate();
+        ProjCoordinate result1 = new ProjCoordinate();
+
+        transformer0.transform(pc, result0);
+        transformer1.transform(pc, result1);
+
+        assertTrue(isInTolerance(result0, 111313.95106842462, 110585.61615828621, 0.000001));
+        assertTrue(isInTolerance(result1, 111313.95105169504, 110585.50558411982, 0.000001));
+    }
+
+    @Test
+    public void epsgWebMercatorLegacyTest() {
+        CRSFactory csFactory = new CRSFactory();
+        try {
+            String code = csFactory.readEpsgFromParameters("+proj=merc +a=6378137 +b=6378137 +lat_ts=0.0 +lon_0=0.0 +x_0=0.0 +y_0=0 +k=1.0 +units=m +nadgrids=@null +wktext +no_defs");
+            Assert.assertEquals(Integer.parseInt(code), 3857);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
 
     private boolean isInTolerance(ProjCoordinate p, double x, double y, double tolerance) {
         /*